add additional future fields in migration
- also update access layer
- add update user and update regression

+++ internal/access/model/user_model.go
@@ -6,7 +6,7 @@ type User struct {
 	Email       string `json:"email"`
 	Username    string `json:"username"`
 	Password    string `json:"password,omitempty"`
-	Rating      uint16 `json:"rating"`
+	Rating      uint32 `json:"rating"`
 	MovesMade   uint64 `json:"moves_made"`
 	GamesPlayed uint32 `json:"games_played"`
 	TimeCreated int64  `json:"time_created"`
